+++ python/qualitylib/domain/__init__.py
@@ -33,6 +33,5 @@ from .software_development.team import Team
 from .software_development.person import Person
 from .software_development.project import Project
 from .software_development.requirement import Requirement
-from .software_development.street import Street
 from .software_development.document import Document
 from .software_development.quality_attribute import QualityAttribute